+++ rust/tableau_summary/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod twb;
+mod xml;
