+++ rust/tableau_summary/src/twb/dashboard.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::twb::NAME_KEY;
 use crate::twb::worksheet::{DataDependencies, get_worksheet_data};
-use crate::twb::xml::XmlExt;
+use crate::xml::XmlExt;
 
 #[derive(Serialize, Deserialize, Default, PartialEq, Clone, Debug)]
 pub struct DashboardMeta {
