+++ rust/tableau_summary/src/twb/datasource/dep.rs
@@ -3,7 +3,7 @@ use roxmltree::Node;
 use serde::{Deserialize, Serialize};
 use tracing::info;
 use crate::twb::datasource::columns::{ColumnDep, get_column_dep_map};
-use crate::twb::xml::XmlExt;
+use crate::xml::XmlExt;
 
 #[derive(Serialize, Deserialize, Default, PartialEq, Clone, Debug)]
 pub struct Dep {
