+++ rust/tableau_summary/src/twb/datasource/object_graph.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use roxmltree::Node;
 use tracing::info;
 use crate::twb::datasource::connection::{Expression, Join, Relation, Table, Union};
-use crate::twb::xml::XmlExt;
+use crate::xml::XmlExt;
 
 #[derive(Serialize, Deserialize, Default, PartialEq, Clone, Debug)]
 pub struct ObjectGraph {
