+++ rust/tableau_summary/src/twb/worksheet.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::twb::datasource::columns::{ColumnDep, get_column_dep_map, GroupFilter};
 use crate::twb::NAME_KEY;
-use crate::twb::xml::XmlExt;
+use crate::xml::XmlExt;
 
 #[derive(Serialize, Deserialize, Default, PartialEq, Clone, Debug)]
 pub struct WorksheetMeta {
